watermarking/requirements.txt
transformers
torch
torchvision
Pillow
scikit-image
imagehash
numpy
opencv-python
fastapi
uvicorn
python-multipart 